+++ electron/ipc/cookie.ts
@@ -1,0 +1,11 @@
+import { ipcMain, session } from "electron";
+
+import { channel } from "./channel";
+
+export function registerCookieIpcHandlers() {
+  ipcMain.handle(channel.cookie.get, async (_, key: string) => {
+    const cookies = await session.defaultSession.cookies.get({ name: key, domain: ".bilibili.com" });
+
+    return cookies?.[0]?.value;
+  });
+}