+++ electron/ipc/index.ts
@@ -1,4 +1,7 @@
+import type { IpcHandlerProps } from "./types";
+
 import { registerAppHandlers } from "./app";
+import { registerCookieIpcHandlers } from "./cookie";
 import { registerDialogHandlers } from "./dialog";
 import { registerDownloadHandlers } from "./download";
 import { registerFontHandlers } from "./font";
@@ -6,12 +9,13 @@ import { registerRequestHandlers } from "./request";
 import { registerStoreHandlers } from "./store";
 import { registerWindowHandlers } from "./window";
 
-export function registerIpcHandlers() {
+export function registerIpcHandlers(props: IpcHandlerProps) {
   registerStoreHandlers();
   registerDialogHandlers();
   registerFontHandlers();
   registerDownloadHandlers();
   registerRequestHandlers();
   registerAppHandlers();
-  registerWindowHandlers();
+  registerCookieIpcHandlers();
+  registerWindowHandlers(props);
 }